Refaktorering av Brev og BrevSkjema. Bruk useBrevApi.
